Move change rate drawer to provider section
