+++ src/renderer/screens/exchange/Swap2/Form/FormSummary/SectionRate.js
@@ -1,25 +1,23 @@
 // @flow
-import React, { useContext, useMemo } from "react";
-import { useSelector } from "react-redux";
-import SummaryLabel from "./SummaryLabel";
-import SummaryValue, { NoValuePlaceholder } from "./SummaryValue";
-import { useTranslation } from "react-i18next";
-import IconLock from "~/renderer/icons/Lock";
-import IconLockOpen from "~/renderer/icons/LockOpen";
-import SummarySection from "./SummarySection";
-import { context } from "~/renderer/drawers/Provider";
-import RatesDrawer from "../RatesDrawer";
 import type {
-  SwapSelectorStateType,
   RatesReducerState,
   SwapDataType,
+  SwapSelectorStateType,
 } from "@ledgerhq/live-common/lib/exchange/swap/hooks";
-import Price from "~/renderer/components/Price";
-import { rateSelector, rateExpirationSelector } from "~/renderer/actions/swap";
-import CountdownTimer from "~/renderer/components/CountdownTimer";
-import Box from "~/renderer/components/Box";
+import React from "react";
+import { useTranslation } from "react-i18next";
+import { useSelector } from "react-redux";
+import { rateExpirationSelector, rateSelector } from "~/renderer/actions/swap";
 import AnimatedCountdown from "~/renderer/components/AnimatedCountdown";
+import Box from "~/renderer/components/Box";
+import CountdownTimer from "~/renderer/components/CountdownTimer";
+import Price from "~/renderer/components/Price";
+import IconLock from "~/renderer/icons/Lock";
+import IconLockOpen from "~/renderer/icons/LockOpen";
 import { ratesExpirationThreshold } from "~/renderer/reducers/swap";
+import SummaryLabel from "./SummaryLabel";
+import SummarySection from "./SummarySection";
+import SummaryValue, { NoValuePlaceholder } from "./SummaryValue";
 
 type Props = {
   ratesState: RatesReducerState,
@@ -30,29 +28,14 @@ type Props = {
 };
 const SectionRate = ({ fromCurrency, toCurrency, ratesState, refetchRates, provider }: Props) => {
   const { t } = useTranslation();
-  const { setDrawer } = useContext(context);
   const exchangeRate = useSelector(rateSelector);
   const ratesExpiration = useSelector(rateExpirationSelector);
-  const rates = ratesState.value;
-  const handleChange = useMemo(
-    () =>
-      rates &&
-      rates.length > 1 &&
-      (() =>
-        setDrawer(RatesDrawer, {
-          fromCurrency,
-          toCurrency,
-          rates,
-          provider,
-        })),
-    [setDrawer, rates, fromCurrency, toCurrency, provider],
-  );
 
   const summaryValue =
     ratesState.status === "loading" ? (
       <NoValuePlaceholder />
     ) : exchangeRate && fromCurrency && toCurrency ? (
-      <SummaryValue handleChange={handleChange}>
+      <SummaryValue>
         {ratesExpiration && exchangeRate.tradeMethod === "fixed" && ratesExpiration > Date.now() && (
           <Box horizontal alignItems="center" mr={2}>
             <Box mr={1}>
